nowe kernele

+++ ConvolutionalNN/ConvolutionalNN/Utils/Utils.cpp
@@ -13,7 +13,7 @@ size_t align(size_t pValue, size_t pAlign){
 }
 
 
-uint bigRand32(){
+uint32 bigRand32(){
 	return
 		((rand() & 0x00FF)			) |
 		((rand() & 0x00FF) << 8		) |
@@ -21,8 +21,8 @@ uint bigRand32(){
 		((rand() & 0xFF00) << 24	);
 }
 
-size_t bigRand64(){
-	return (static_cast<size_t>(bigRand32()) << 32) | static_cast<size_t>(bigRand32());
+uint64 bigRand64(){
+	return (static_cast<uint64>(bigRand32()) << 32) | static_cast<uint64>(bigRand32());
 }
