+++ ConvolutionalNN/ConvolutionalNN/Utils/Utils.h
@@ -11,8 +11,8 @@ namespace cnn {
 
 size_t align(size_t pValue, size_t pAlign);
 
-uint	bigRand32();
-size_t	bigRand64();
+uint32 bigRand32();
+uint64 bigRand64();
 
 
 	}
